suggestions for refactoring and new architecture
