+++ passwordWindow.js
@@ -1,21 +1,13 @@
+import {hashPassword} from "./src/crypt.js";
+
 document.querySelector('.passwordWindowForm').addEventListener('submit', (e) => {
   e.preventDefault();
   const password = document.getElementById('passwordInputField').value;
 
-  // Initialize SJCL's random number generator
-  sjcl.random.startCollectors();
-
-  // Generate a random salt
-  const saltBits = sjcl.random.randomWords(2); // Generates a 64-bit salt
-  const saltHex = sjcl.codec.hex.fromBits(saltBits);
-
-  // Use PBKDF2 to hash the password with the salt
-  const iterations = 10000;
-  const hashedPasswordBits = sjcl.misc.pbkdf2(password, saltBits, iterations, 256);
-  const hashedPasswordHex = sjcl.codec.hex.fromBits(hashedPasswordBits);
+  let [hashedPassword, salt ] = hashPassword(password);
 
   // Prepare data to send
-  const data = { hashedPassword: hashedPasswordHex, salt: saltHex };
+  const data = { hashedPassword: hashedPassword, salt: salt };
 
   // Use the exposed API to send the hashed password and salt
   window.electronAPI.submitPassword(data);
