+++ src/pages/main/index.html
@@ -8,8 +8,8 @@
 
     />
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="node_modules/bootstrap/dist/css/bootstrap.min.css">
-    <link href="styles.css" type="text/css" rel="stylesheet">
+    <link rel="stylesheet" href="../../../node_modules/bootstrap/dist/css/bootstrap.min.css">
+    <link href="../../../styles.css" type="text/css" rel="stylesheet">
     <title>Password Frog</title>
 </head>
 <body>
@@ -20,7 +20,7 @@
             </div>
             <button type="submit" class="btn btn-primary mt-2">Submit</button>
         </form>
-    </div>    
+    </div>
 
 </body>
 </html>
