+++ src/pages/password-input/passwordInput.html
@@ -8,9 +8,8 @@
     >
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>Enter Password</title>
-    <link rel="stylesheet" href="node_modules/bootstrap/dist/css/bootstrap.min.css">
-    <link href="styles.css" type="text/css" rel="stylesheet">
-    <script src="sjcl.js"></script>
+    <link rel="stylesheet" href="../../../node_modules/bootstrap/dist/css/bootstrap.min.css">
+    <link href="../../../styles.css" type="text/css" rel="stylesheet">
 </head>
 <body>
     <div class="passwordWindowWrapper">
